Move base classes into base package
